save fetched lyrics to id3tags of the mp3 file
Fetched lyrics can be automatically saved to id3tags of the mp3 file.
To enable this feature eyeD3 python module should be properly installed
and the corresponding option in preferences window (on display tab)
should be checked.

Lyrics are save in 'utf-8' encoding, default eyeD3 language (English)
and 'sonata' as the description. If the mp3 file already contains lyrics
tag with the same language and description it's not overwritten.

Add '(save)' button to info tab to manually save lyrics to id3tags
(overwrite existing lyrics with the same language and description).