main: pass proper widget to tray tooltip on song change
Fix jumping tray icon tooltip for a moment from its position near tray
icon  to the (0, 0) position on the screen if gtk.StatusIcon is used
and song is changed while tooltip is visible (mouse is over tray icon).
That is caused by passing self.trayeventbox (regardless of tray icon
type currently used), that is None if we use gtk.StatusIcon, to
self.traytips._real_display method.